pyproject.toml
[tool.poetry]
name = "gearstore"
version = "0.1.0"
description = ""
authors = ["Gleb Ryabets"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.21"
pydantic-settings = "^2.0.3"
psycopg2 = "^2.9.9"
alembic = "^1.12.0"
click = "^8.1.7"
redis = "^5.0.1"
bcrypt = "^4.0.1"
email-validator = "^2.0.0.post2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
